Add a section about fixups to pull-requests
It has been asked about elsewhere, so it's as well to be explicit here.
Fixup commits are a little problematic in that their commit messages are
unrelated to the commit contents, so they're hard for a reviewer to
parse without doing some detective work. It makes life easier for the
reviewer to squash them before opening a PR.